+++ src/titiler/mosaic/tests/test_factory.py
@@ -15,6 +15,7 @@ from cogeo_mosaic.backends import FileBackend
 from cogeo_mosaic.backends import MosaicBackend as MosaicJSONBackend
 from cogeo_mosaic.mosaic import MosaicJSON
 from fastapi import FastAPI, Query
+from owslib.wmts import WebMapTileService
 from rio_tiler.mosaic.methods import PixelSelectionMethod
 from starlette.testclient import TestClient
 
@@ -796,3 +797,55 @@ def test_ogc_maps_mosaic():
             },
         )
         assert response.status_code == 422
+
+
+def test_wmts_extension_mosaic():
+    """Test wmtsExtension extension."""
+    mosaic = MosaicTilerFactory(
+        backend=MosaicJSONBackend,
+        extensions=[
+            wmtsExtension(
+                get_renders=lambda obj: {
+                    "one_band_limit": {
+                        "tilematrixsets": {
+                            "WebMercatorQuad": [0, 1],
+                        },
+                        "spatial_extent": (-180, -90, 180, 90),
+                        "bidx": [1, 2, 3],
+                        "rescale": [[0, 10000], [0, 5000], [0, 1000]],
+                    },
+                    "one_band": {
+                        "bidx": [1],
+                        "rescale": ["0,10000"],
+                    },
+                },
+            )
+        ],
+    )
+    app = FastAPI()
+    app.include_router(mosaic.router)
+    add_exception_handlers(app, MOSAIC_STATUS_CODES)
+
+    with TestClient(app) as client:
+        with tmpmosaic() as mosaic_file:
+            response = client.get(f"/WMTSCapabilities.xml?url={mosaic_file}")
+            assert response.status_code == 200
+            assert response.headers["content-type"] == "application/xml"
+            meta = parse_img(response.content)
+            assert meta["driver"] == "WMTS"
+
+            wmts = WebMapTileService(
+                f"/WMTSCapabilities.xml?url={mosaic_file}", xml=response.content
+            )
+            assert wmts.version == "1.0.0"
+            assert len(wmts.contents) == 39  # (2 renders + default) x 13 TMS
+
+            layer = wmts.contents[f"{mosaic_file}_WebMercatorQuad_default"]
+            assert ["7", "8", "9"] == list(
+                layer.tilematrixsetlinks["WebMercatorQuad"].tilematrixlimits
+            )
+
+            layer = wmts.contents[f"{mosaic_file}_WebMercatorQuad_one_band_limit"]
+            assert ["0", "1"] == list(
+                layer.tilematrixsetlinks["WebMercatorQuad"].tilematrixlimits
+            )